refactor(ssh): split session package into multiple files
